9 sep commit
Not complete with exceptions
