+++ src/main/java/exception/InvalidArgumentsException.java
@@ -1,0 +1,4 @@
+package exception;
+
+public class InvalidArgumentsException extends Exception{
+}
