+++ src/main/java/exception/InvalidCommandException.java
@@ -1,0 +1,4 @@
+package exception;
+
+public class InvalidCommandException extends Exception{
+}
