+++ src/main/java/exception/TaskDoesNotExistException.java
@@ -1,0 +1,4 @@
+package exception;
+
+public class TaskDoesNotExistException extends Exception{
+}
