+++ src/main/java/task/Deadline.java
@@ -1,16 +1,15 @@
-public class Event extends Task{
+package task;
 
+public class Deadline extends Task {
     protected String by;
-
-    public Event(String description, String by) {
+    public Deadline(String description, String by) {
         super(description);
         this.by = by;
         System.out.println("Got it. I've added this task:");
         System.out.println("  " + toString());
     }
-
     @Override
     public String toString() {
-        return ("[E][" + super.getStatusIcon() + "] " + super.description + " (at: " + by + ")");
+        return ("[D][" + super.getStatusIcon() + "] " + super.description + " (by: " + by + ")");
     }
 }
