+++ src/main/java/task/Event.java
@@ -1,13 +1,18 @@
-public class Deadline extends Task{
+package task;
+
+public class Event extends Task{
+
     protected String by;
-    public Deadline(String description, String by) {
+
+    public Event(String description, String by) {
         super(description);
         this.by = by;
         System.out.println("Got it. I've added this task:");
         System.out.println("  " + toString());
     }
+
     @Override
     public String toString() {
-        return ("[D][" + super.getStatusIcon() + "] " + super.description + " (by: " + by + ")");
+        return ("[E][" + super.getStatusIcon() + "] " + super.description + " (at: " + by + ")");
     }
 }
