+++ src/main/java/task/TaskList.java
@@ -1,3 +1,5 @@
+package task;
+
 public class TaskList {
     public static final String COMMAND_MARKED = "mark";
     public static final String COMMAND_TODO = "todo";
